Added basic framework to run webdriver chrome tests locally
The dependencies for selenium, testng and lo4j2 are added in the framework.
A testng listener implementation is also added, along with the retry analyzer
implementation. The number of retires are configured in a config file.
The webdriver creation is made threadsafe so that the tests can be run in
parallel. Only one test class with a test is added, to ensure it can be tested.